backend/requirements.txt
aiofiles==23.2.1
alembic==1.13.1
aniso8601==9.0.1
asgiref==3.8.1
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-CLI==0.4.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
flask-cors==4.0.0